# Fix tx buffer inconsistency if there are unordered key writes in one tx.