@fuweid 2024-01-17
Hi @siyuanfoundation you can try to limit cpu by the following command and use patch #17248

```
EXPECT_DEBUG=true GO_TEST_FLAGS=${GO_TEST_FLAGS} RESULTS_DIR=/tmp/results taskset -c 0,1,2 make test-robustness
```

Hope it can help.