@serathius 2024-03-27
> Here are the results: (they are basically the same)

Benchstat do not change the results, it just makes them readable and makes it easy to check if they are [statistically significant ](https://en.wikipedia.org/wiki/Statistical_significance)

If you provide both old and new benchmark results to benchstat, it will provide also the performance delta and [p-value](https://en.wikipedia.org/wiki/P-value)